refactor(internal/pkg/document): consolidate routes into single file
